Remove dependency intl-parse-accept-language
